requirements.txt
behave
selenium
webdriver-manager
allure-behave
python-dotenv
requests